feat: change sophystory to predicted booktalk
