+++ apis/sophyStory.ts
@@ -2,9 +2,8 @@ import Cookies from 'js-cookie';
 import api from '.';
 
 export const fetchSophyStory = async () => {
-  const memberId = Cookies.get('memberId');
   const {
     data: { data },
-  } = await api.get(`/sophy-story/${memberId}/all`);
+  } = await api.get(`/sophy-story/booktalks`);
   return data;
 };
